--- .rubocop.yml
@@ -1,2 +1,0 @@
-AllCops:
-  TargetRubyVersion: 2.5
